+++ project/src/components/app/app.tsx
@@ -7,7 +7,7 @@ import { getComments } from '../../mock/comments';
 import { PropertyScreen } from '../../pages/property-screen/property-screen';
 import { Hotel } from '../../types/hotel';
 import { User } from '../../types/user';
-import { AppRoute, AuthorizationStatus } from '../../const/const';
+import { AppRoute, AuthorizationStatus } from '../../consts/consts';
 import { NotFoundScreen } from '../../pages/not-found-screen/not-found-screen';
 import { PrivateRoute } from '../private-route/pravate-route';
 
@@ -48,7 +48,7 @@ const App: React.FunctionComponent<MainScreenProps> = ({ currentCity, currentSor
         }
       />
       <Route
-        path={`${AppRoute.Room}/:id`}
+        path={`${AppRoute.Room}`}
         element={
           < PropertyScreen
             hotelId={1}
