+++ project/src/components/place-card/place-card.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { Hotel } from '../../types/hotel';
-import { cityCardType, hotelType } from '../../const/const';
+import { cityCardType, hotelType } from '../../consts/consts';
 
 type PlaceCardProps = {
   hotel: Hotel;
@@ -9,9 +9,7 @@ type PlaceCardProps = {
 
 export const PlaceCard: React.FunctionComponent<PlaceCardProps> = ({ hotel, cardType }) => (
   <article className={`${cardType} place-card`}>
-    {hotel.isPremium
-      ? <div className="place-card__mark"> <span>Premium</span> </div>
-      : ''}
+    {hotel.isPremium && <div className="place-card__mark"> <span>Premium</span> </div>}
     {cardType === cityCardType.CITIES_CARD &&
       (
         <div className="cities__image-wrapper place-card__image-wrapper">
@@ -39,7 +37,8 @@ export const PlaceCard: React.FunctionComponent<PlaceCardProps> = ({ hotel, card
         </div>
         <button className={`place-card__bookmark-button button ${hotel.isFavorite
           ? 'place-card__bookmark-button--active'
-          : ''}`} type="button"
+          : ''}`}
+        type="button"
         >
           <svg className="place-card__bookmark-icon" width="18" height="19">
             <use xlinkHref="#icon-bookmark"></use>
