+++ project/src/consts/consts.ts
@@ -23,7 +23,7 @@ export enum AppRoute {
 Main = '/',
 Login = '/login',
 Favorites = '/favorites',
-Room = '/offer'
+Room = '/offer/:id'
 }
 
 export enum AuthorizationStatus {
