+++ project/src/mock/hotels.ts
@@ -5591,8 +5591,8 @@ const hotels: Hotel[] =
 
 export const getHotels = () => hotels;
 
-export const getHotelsByCity = (city: string): Hotel[] => hotels.filter((item: Hotel) => item.city.name === city);
+export const getHotelsByCity = (city: string) => hotels.filter((item: Hotel) => item.city.name === city);
 
 export const getHotelById = (id: number) => hotels.find((item: Hotel) => item.id === id);
 
-export const getFavoriteHotels = (): Hotel[] => hotels.filter((item: Hotel) => item.isFavorite);
+export const getFavoriteHotels = () => hotels.filter((item: Hotel) => item.isFavorite);
